Test Backend: n_shots in get_counts and expect...
